refactor: isolate chat event index reads
